+++ wall_factory_2.py
@@ -1,4 +1,189 @@
+import math
+
+from mathutils import Vector
+
+from board_factory import BoardFactory
+from windows import WindowFactory
+
+
 class WallFactory:
+    @staticmethod
+    def create_framed_wall(
+            name,
+            length,
+            height,  # total wall height, including bottom + top plates
+            location=(0, 0, 0),
+            stud_spacing=STUD_SPACING,
+            stud_width=STUD_WIDTH,
+            bottom_plate_count=1,
+            top_plate_count=2,
+            add_sheathing=True,
+            add_drywall=True,
+            window_specs=None,
+            materials=None,
+            stud_spec="2x4",
+    ):
+        material_framing = materials['framing']
+        material_sheathing = materials['sheathing']
+        material_drywall = materials['drywall']
+        material_glass = materials['glass']
+        """
+        Creates a rectangular framed wall:
+          - bottom_plate_count (1 typical),
+          - top_plate_count (2 typical, for double top plates),
+          - vertical studs spaced on 'stud_spacing',
+          - optional single window opening,
+          - optional exterior sheathing (one big panel),
+          - optional interior drywall (one big panel).
+
+        Returns a Collection containing all parts of this wall.
+        """
+        stud_thickness = WallFactory.STUD_SPECS[stud_spec]["thickness"]  # Always 1.5"
+        stud_width = WallFactory.STUD_SPECS[stud_spec]["width"]  # 3.5", 5.5", etc.
+
+        # **STEP 1: Create the Parent Wall Object**
+        bpy.ops.object.empty_add(type='PLAIN_AXES', location=location)
+        wall_parent = bpy.context.object
+        wall_parent.name = name
+
+        """
+        Ensures the wall's origin is at its correct center.
+        """
+        bpy.ops.object.select_all(action='DESELECT')
+        wall_parent.select_set(True)
+        bpy.context.view_layer.objects.active = wall_parent
+
+        # Set origin to the bottom center
+        bpy.ops.object.origin_set(type='ORIGIN_GEOMETRY', center='BOUNDS')
+
+        # Apply all transforms (scale, rotation, location)
+        bpy.ops.object.transform_apply(location=False, rotation=True, scale=True)
+
+        # 1) BOTTOM PLATES
+        plate_thickness = stud_thickness
+        plate_depth = stud_width
+        # wall_parent.matrix_world.inverted() @ plate_world_position
+        total_bottom_height = plate_thickness * bottom_plate_count
+        for i in range(bottom_plate_count):
+            z_plate = plate_thickness / 2.0 + i * plate_thickness
+            plate_position = Vector((0, 0, z_plate))
+            # WallFactory.create_child(wall_parent,f"{name}_BottomPlate_{i + 1}",length,0)
+            plate = BoardFactory.add_board(wall_parent,
+                                           board_name=f"{name}_BottomPlate_{i + 1}",
+                                           length=length,
+                                           height=plate_thickness,
+                                           depth=plate_depth,
+                                           location=plate_position,
+                                           material=material_framing
+                                           )
+            # plate.parent = wall_parent
+            # plate.location = wall_parent.matrix_world.inverted() @ plate_world_position
+            # move_to_collection(plate, wall_coll)
+
+        # 2) TOP PLATES
+        total_top_height = plate_thickness * top_plate_count
+        z_top_start = height - total_top_height
+        for i in range(top_plate_count):
+            z_plate = z_top_start + plate_thickness / 2.0 + i * plate_thickness
+            plate_position = Vector((0, 0, z_plate))
+            plate = BoardFactory.add_board(wall_parent,
+                                           board_name=f"{name}_TopPlate_{i + 1}",
+                                           length=length,
+                                           height=plate_thickness,
+                                           depth=stud_width,
+                                           location=plate_position,
+                                           material=material_framing
+                                           )
+            # plate.parent = wall_parent  # Parent to the main wall object
+            # plate.location = wall_parent.matrix_world.inverted() @ plate_world_position
+
+        # 3) STUD REGION
+        # The "net" region for studs is from top of bottom plates to bottom of top plates.
+        stud_region_height = height - (total_bottom_height + total_top_height)
+        if stud_region_height < 0:
+            stud_region_height = 0
+        z_stud_bottom = total_bottom_height
+
+        # Place studs along X at 'stud_spacing' intervals
+        x_left = - (length / 2.0) + (stud_thickness / 2.0)
+        n_studs = int((length - stud_width) // stud_spacing) + 1
+
+        for i in range(n_studs):
+            x_i = x_left + i * stud_spacing
+            if x_i + stud_width / 2.0 > (length / 2.0):
+                break
+            z_center = z_stud_bottom + stud_region_height / 2.0
+            stud_position = Vector((x_i, 0, z_center))
+            stud_obj = BoardFactory.add_board(wall_parent,
+                                              board_name=f"{name}_Stud_{i + 1}",
+                                              length=stud_thickness,
+                                              height=stud_region_height,
+                                              depth=stud_width,
+                                              location=stud_position,
+                                              material=material_framing
+                                              )
+            # move_to_collection(stud_obj, wall_coll)
+            # stud_obj.parent = wall_parent
+            # stud_obj.location = wall_parent.matrix_world.inverted() @  stud_world_position
+
+        # 5) Exterior Sheathing
+        if add_sheathing:
+            sheathing_thickness = 0.012  # ~12 mm
+            # Sheathing goes on the "outside" (we'll treat negative Y as outside)
+            sheathing_y = - (plate_depth / 2.0 + sheathing_thickness / 2.0)
+
+            BoardFactory.add_cladding(
+                wall_name=name,
+                cladding_type="Sheathing",
+                wall_length=length,
+                wall_height=height,
+                thickness=sheathing_thickness,
+                y_offset=sheathing_y,
+                material=material_sheathing,
+                location=location,
+                wall=wall_parent,
+                sheet_length=2.44,  # Standard 8ft sheet
+                sheet_height=1.22  # Standard 4ft sheet
+            )
+
+            sheathing_z = height / 2.0
+
+        # 6) Interior Drywall
+        if add_drywall:
+            drywall_thickness = 0.0127  # 1/2 inch ~ 0.0127 m
+            # We'll treat +Y as inside, so place drywall behind studs
+            drywall_y = (plate_depth / 2.0 + drywall_thickness / 2.0)
+            drywall_z = height / 2.0
+            drywall_obj = BoardFactory.add_board(wall_parent,
+                                                 board_name=f"{name}_Drywall",
+                                                 length=length,
+                                                 height=height,
+                                                 depth=drywall_thickness,
+                                                 location=(0, drywall_y, drywall_z),
+                                                 material=material_drywall
+                                                 )
+            # move_to_collection(drywall_obj, wall_coll)
+            drywall_obj.parent = wall_parent
+
+        # 4) WINDOW FRAMING (optional)
+        if window_specs:
+            for ws in window_specs:
+                WindowFactory.create_window_opening(
+                    wall=wall_parent,
+                    name_prefix=f"{name}_Window",
+                    window_center_x=ws["center_x"],
+                    window_bottom_z=ws["bottom_z"],
+                    window_width=ws["width"],
+                    window_height=ws["height"],
+                    bottom_plate_height=total_bottom_height,
+                    top_plate_height=plate_thickness,
+                    stud_spec=spec,
+                    wall_height=height,
+                    second_top_plate_height=plate_thickness if top_plate_count > 1 else 0,
+                    material=material_framing,
+                    glass_material=material_glass
+                )
+        return wall_parent
     @staticmethod
     def create_child(parent, board_name, length, height, depth, world_location, material):
         """
